Move Win32::Registry#write monkeypatch into win32/registry monkeypatch file.
